appended convert_poly.py, multi_cheb.py, and multi_power.py to polynomial.py. Fixed import statements to reflect changes. Removed files that were appended.
