+++ groebner/Macaulay.py
@@ -1,17 +1,18 @@
+# Dependencies
 from operator import itemgetter
 import itertools
 import numpy as np
-from groebner import maxheap
 import math
-from groebner.multi_cheb import MultiCheb
-from groebner.multi_power import MultiPower
-from groebner.polynomial import Polynomial
 from scipy.linalg import lu, qr, solve_triangular
 from scipy.sparse import csc_matrix, vstack
-from groebner.maxheap import Term
 import matplotlib.pyplot as plt
 import time
 from collections import defaultdict
+
+# groebner module imports
+from groebner.utils import Term
+from groebner.polynomial import Polynomial, MultiCheb, MultiPower
+
 
 def Macaulay(initial_poly_list, global_accuracy = 1.e-10):
     """
